[#155714607] Adds jasmine tests for canceling preferences and preference opt outs
Because of the change in
https://github.com/Exygy/sf-dahlia-web/commit/bed4b750102fe58f5c33dc4f3e4698fb6d8d322c#diff-bf948082a3b5ee08694d1e3427c28a7fL530
preferences that weren't being claimed were still being reset. This had a side effect of
cancelling claiming and opting out of live/work preference, because NRHP was always being
reset. When applicant is eligible for NRHP, this is the correct behavior, because claiming
NRHP also claims Live/Work.

This commit adds ShortFormApplicationService tests for these scenarios.