Bugs/Live-Work opt out canceled incorrectly #155714607 (#1009)
Because of the change in bed4b75 preferences that weren't being claimed were still being reset. This had a side effect of cancelling the claiming and opting out of Live/Work preference, because NRHP was always being reset before application was sent or loaded from Salesforce.

When applicant is eligible for NRHP, this is the correct behavior, because claiming NRHP also claims Live/Work, if changing address resets NRHP it should also reset Live/Work, etc. However if applicant isn't eligible for NRHP, we don't want NRHP affecting anything with Live/Work preference.

Also, ADHP resets Live/Work in the same way that NRHP does, so additional conditions are now in place to make NRHP/ADHP function in the same way.

* [#155714607] Fixes Live/Work opt out being canceled

* [#155714607] Prevents NRHP/ADHP from incorrectly cancelling Live/Work

* [#155714607] Adds jasmine tests for canceling preferences and preference opt outs