[#154187409] Stops the move from the app confirmation page to itself from being treated as going back (#994)
This correction was brought about because we were seeing a bug wherein using the language switcher from the app confirmation page would send the user out of the application, instead of re-displaying the app confirmation page in the selected language. This was happening because when the language switcher tried to move from the confirmation page state (in the original language) to the confirmation page state (in the selected language), the hittingBackFromConfirmation function viewed this as an attempt to go back from the confirmation page into a previous app page, and would redirect the user out of the application. The correction in this commit now allows the correct language switching behavior to happen on the app confirmation page.