[#150870824] Corrects redirect behavior for confirmation and review submitted pages
The default redirect behavior for when a user tries to access a short form page they don't have access to is to redirect them to the beginning of the application. This commit extends that behavior to the short form confirmation and review submitted pages, which previously were not covered by the logic that implemented the behavior for the rest of the short form pages.