[#156284730] Refactors timing of Google Translation
We previously were beginning Google Translation in the state change start callback. This was occasionally causing the GTranslate process to partially or wholly fail in instances where the page content did not load in the right time or with the right timing for GTranslate to pick up on the content. Moving the kickoff of Google Translation into the state change success callback gives the page content more time to load so that GTranslate has a better chance to pick up on all the content.