[#155955218] Prevents re-triggering of $digest loop from $timeout when not needed
All of our current uses of $timeout are related to waiting for DOM elements on the page
to be rendered and do not need to re-trigger the $digest loop to recalculate data. By
changing the third argument of $timeout to false, we prevent $digest loop from starting again
after the callback function runs.