Bugs/scroll position maintained on home page #156241951
This bug was being caused by two specific conditions:

1. The home page url was specified as '/'. When combined with a language param, this required
the home page url to have a trailing slash, like '/es/', unlike any other routes. To link
back to the non-english homepage, used ng-href and appending a trailing slash to the route
url.

2. The keepScrollPosition directive uses a difference between $location.path() at the beginning
and end of state changes to detect if the browser back/forward buttons were in use or not. When
navigating to the home page, it treated that navigation as if the browser buttons were being
used, and so maintained scroll position for the page.

To solve this:
- Homepage url is changed from '/' to '' to not require the trailing slash for non-english home page
- `ui-state` directive is used to dynamically generate the links for the language switcher
- Logo home page link uses ui-sref with route name rather than a custom function

Also note that angular-ui-router was updated to 0.4.2 (most recent version without breaking
changes) to use the `ui-state` directive that let's us dynamically generate a state link