[#155640979] Refactoring to reduce cognitive complexity of geocoding_data
Code Climate rated cognitive complexity of GeocodingController#geocoding_data at 9, which is over our set limit of 8, so this refactor is to reduce that. Note that a lot of the logging in GeocodingController is of a temporary nature to help illuminate some issues we've been seeing around geocoding/boundary matching. The logging will likely be removed after that issue is resolved.