Chores/google translate bug stopgap #156934838 (#1032)
* [#156284730] Refactors timing of Google Translation

We previously were beginning Google Translation in the state change start callback. This was occasionally causing the GTranslate process to partially or wholly fail in instances where the page content did not load in the right time or with the right timing for GTranslate to pick up on the content. Moving the kickoff of Google Translation into the state change success callback gives the page content more time to load so that GTranslate has a better chance to pick up on all the content.

* [#155955218] Defer automated translation with $timeout

Use built-in $timeout rather than standard setTimeout.

* [#155955218] Prevent $digest loop from restarting while waiting for automated translation

We do not need to re-trigger a $digest loop while waiting for translation functionality to be
available, because this process does not change any internal application data.

* [#155955218] Prevents re-triggering of $digest loop from $timeout when not needed

All of our current uses of $timeout are related to waiting for DOM elements on the page
to be rendered and do not need to re-trigger the $digest loop to recalculate data. By
changing the third argument of $timeout to false, we prevent $digest loop from starting again
after the callback function runs.

* [#155955218] WIP - adds tests for ExternalTranslateService

* [#156934838] Adds quick fix for Google Translate timing issue on listing/s pages

This is meant to be a temp fix; we have another story in the backlog to implement a more thorough fix.