Features/geocoder project id adhp #155640979 (#1034)
* [#155640979] Makes boundary matching use ADHP as project ID when listing has ADHP

* [#155640979] Adds Jasmine spec for GeocodingService

* [#155640979] Removes unused NRHP value from geocoding options

* [#155640979] Adds a check for ADHP or NRHP before boundary match attempt

Adds more rspec tests for the various geocoding and boundary matching scenarios. Updates the Jasmine tests for geocoding and boundary matching.

* [#155640979] Refactoring to reduce cognitive complexity of geocoding_data

Code Climate rated cognitive complexity of GeocodingController#geocoding_data at 9, which is over our set limit of 8, so this refactor is to reduce that. Note that a lot of the logging in GeocodingController is of a temporary nature to help illuminate some issues we've been seeing around geocoding/boundary matching. The logging will likely be removed after that issue is resolved.

* [#155640979] Refactors geocoding and boundary matching

This refactor reduces mixing of concerns in the handling of preferences, geocoding, and boundary matching. Geocoding and boundary matching are more clearly indicated as two separate but related processes. All handling of preference data is removed from the GISService so that it now only deals with GIS data.

* [#155640979] Slight additional refactor of GIS data promise usage

* [#155640979] Refactors the short form address validations; adds Jasmine tests for afterGeocode

* [#155640979] Small corrections to GISService and related tests