[#162065874] Updates Coffeelint configuration used by Overcommit
The previous Coffeelint configuration in overcommit.yml was not working for me - Overcommit would throw an error when I ran it and it reached the stage where it tried to run Coffeelint. I used the information in this Overcommit GitHub issue: https://github.com/brigade/overcommit/issues/352, to update the configuration to something that worked without error.