[#162494687] Return to use of global vars for env vars in Angular app
The solutions we've tried so far for using Angular constants to provide environment variables to the Angular app have not been compatible with running in production on Heroku. We're returning to using global variables for now. We've created a follow-up story to implement a method of providing env vars in Angular that does not use global vars.