[#165609939] Upgrade the jasmine gem to 3.4.0 and begin fixing failing tests
In Jasmine 3, it is now the default to run tests in random order. Previously, we were not running tests in random order. Running them randomly has exposed many tests where the test setup was not isolated enough and so the test failed if certain other tests were run before. In this story, we'll be fixing all those newly failing tests after the Jasmine upgrade.