[#167549021] Refactor review application details partial into component
As part of moving towards a more component-based architecture, both to make working in the current codebase easier and to make the eventual future upgrade of the codebase easier, we are starting to componentize parts of the short form application. We aim to reduce usage of scope, to move code out of the much-too-large ShortFormApplicationService and ShortFormApplicationController, and to gather related functionalities together so we can better see what is used where. As a step in those directions, this commit converts the review-application-details partial into the applicationSummary component. In future, this applicationSummary component could be replaced by smaller components, perhaps one for each section of the summary.