Add qualify to app review #167549021 (#1253)
* [#167549021] First pass at getting qualify section into application review

* [#167549021] Add tests for lending agent retrieval

* [#167549021] Refactor lending agent functions

* [#167549021] Only show qualify section on sale applications

* [#167549021] Refactor review application details partial into component

As part of moving towards a more component-based architecture, both to make working in the current codebase easier and to make the eventual future upgrade of the codebase easier, we are starting to componentize parts of the short form application. We aim to reduce usage of scope, to move code out of the much-too-large ShortFormApplicationService and ShortFormApplicationController, and to gather related functionalities together so we can better see what is used where. As a step in those directions, this commit converts the review-application-details partial into the applicationSummary component. In future, this applicationSummary component could be replaced by smaller components, perhaps one for each section of the summary.

* [#167549021] Add tests for preference summary component

Move tests from summary tests, get a few tests working, add tests for fileAttachmentsForRentBurden

* [#167937039] Add additional application summary tests

Add tests for applicationIncomeAmount, continueDraftApplicantHasUpdatedInfo

* [#167549021] Move all qualify review content to locale-en

* [#167549021] Run grunt translations

* [#167549021] Move file attachment label generation to component