Upgrade devise_token_auth to fix emails #173504864
devise_token_auth overrides some methods that devise uses to send
emails. See this issue: https://github.com/heartcombo/devise/issues/3654
for more information.

It was fixed by this pr:
https://github.com/lynndylanhurley/devise_token_auth/pull/1343

In release 1.1.4 of devise_token_auth. This change updates that library
to include the fix in webapp.

Additionally, the devise library appears to have an issue in 4.7.2, so
I explicitly downgraded us to 4.7.1.

We are able to get rid of the custom email triggering code with this
fix.